get client ip addr via websocket call
We can use a mix of dualstack / v4-only / v6-only URLs to
request both the client ipv4-address and the ipv6-address.

But, with HTTP/2 connections can be re-used, and because
some of these URLs can share addresses, an open dual-stack
connection might be reused for a connection to the v6-only
URL.

So for now we just request the client IP address(es) via
a websocket connection since there is no re-use for
other vhosts in that case.